Also moved useInfiniteScroll out
